requirements.txt
Flask==1.1.2
gunicorn==20.0.4
Pillow==8.1.0
python-dotenv==0.15.0
requests==2.25.1